[typescript] Change: Export type directly, add index.ts (move from index.d.ts)
