--- index.js
@@ -1,7 +1,0 @@
-
-import PhotoManipulator from './lib/PhotoManipulator';
-
-export { PhotoManipulator };
-
-
-export default PhotoManipulator;
