+++ package.json
@@ -2,8 +2,9 @@
   "name": "react-native-photo-manipulator",
   "version": "1.3.1",
   "description": "React Native Image Processing API to edit photo programmatically for Android and iOS.",
-  "main": "index.js",
-  "typings": "index.d.ts",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
+  "source": "src/",
   "author": "Woraphot Chokratanasombat <guhungry@gmail.com>",
   "scripts": {
     "ci": "CI=true && yarn lint && yarn test",
@@ -56,8 +57,6 @@
     "android",
     "ios",
     "lib",
-    "index.js",
-    "index.d.ts",
     "react-native-photo-manipulator.podspec",
     "!/**/__tests__"
   ],
