+++ src/index.ts
@@ -1,0 +1,7 @@
+import PhotoManipulator from './PhotoManipulator';
+
+// Type
+export * from './PhotoManipulatorTypes';
+
+export {PhotoManipulator};
+export default PhotoManipulator;
